backend/requirements.txt
# FastAPI Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0

# Database Clients
qdrant-client>=1.7.0
psycopg[binary]>=3.1.0
redis>=5.0.0

# AI/ML Libraries
groq>=0.4.0
sentence-transformers>=2.2.0

# Pydantic for validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code Quality
black>=23.11.0
ruff>=0.1.0
mypy>=1.7.0
